chore: Split functions into different files and make loadable
