+++ nodejs/genai/gemini/chatbot/src/gitFunctions.js
@@ -498,171 +498,9 @@ async function listGitHubActions(username, repoName, status = 'in_progress') {
 
 /* eslint-enable no-restricted-syntax, no-await-in-loop, consistent-return */
 
-/** Registry of available GitHub functions for AI. */
-const availableFunctionsRegistry = {
-  create_pull_request: {
-    func: createGithubPullRequest,
-    params: ['username', 'repoName', 'title', 'sourceBranch', 'targetBranch', 'body'],
-  },
-  fetch_repo_contents: {
-    func: fetchRepoContentsRecursive,
-    params: ['username', 'repoName', 'repoPath', 'localDestPath', 'includeDotGithub'],
-  },
-  list_actions: {
-    func: listGitHubActions,
-    params: ['username', 'repoName', 'status'],
-  },
-  list_public_repos: {
-    func: listPublicRepos,
-    params: ['username'],
-  },
-  list_branches: {
-    func: listBranches,
-    params: ['username', 'repoName'],
-  },
-  list_commit_history: {
-    func: listCommitHistory,
-    params: ['username', 'repoName', 'filePath'],
-  },
-  list_directory_contents: {
-    func: listDirectoryContents,
-    params: ['username', 'repoName', 'path'],
-  },
-};
-
-/** Metadata for GitHub functions as AI tools. */
-const funcs = [
-  {
-    type: 'function',
-    function: {
-      name: 'create_pull_request',
-      description: 'Create a pull request on a given GitHub repository.',
-      parameters: {
-        type: 'object',
-        properties: {
-          username: { type: 'string', description: 'The GitHub username.' },
-          repoName: { type: 'string', description: 'The repository name.' },
-          title: { type: 'string', description: 'The Pull Request title.' },
-          sourceBranch: { type: 'string', description: 'The source branch name.' },
-          targetBranch: { type: 'string', description: 'The target branch name.' },
-          body: { type: 'string', description: 'The description or body of the pull request.' },
-        },
-        required: ['username', 'repoName', 'title', 'sourceBranch', 'targetBranch'],
-      },
-    },
-  },
-  {
-    type: 'function',
-    function: {
-      name: 'fetch_repo_contents',
-      description: 'Fetch or download the contents of a GitHub repository.',
-      parameters: {
-        type: 'object',
-        properties: {
-          username: { type: 'string', description: 'The GitHub username.' },
-          repoName: { type: 'string', description: 'The repository name.' },
-          repoPath: { type: 'string', description: 'The GitHub repository path to start download at.' },
-          localDestPath: { type: 'string', description: 'The target local directory path to download to.' },
-          includeDotGithub: { type: 'boolean', description: 'A boolean to include .github metadata or not.' },
-          retryCount: { type: 'number', description: 'The retry count to use.' },
-          maxRetries: { type: 'number', description: 'The maximum number of retries.' },
-        },
-        required: ['username', 'repoName', 'repoPath', 'localDestPath', 'includeDotGithub', 'retryCount', 'maxRetries'],
-      },
-    },
-  },
-  {
-    type: 'function',
-    function: {
-      name: 'list_public_repos',
-      description: 'Lists public repositories for a given GitHub username.',
-      parameters: {
-        type: 'object',
-        properties: {
-          username: { type: 'string', description: 'The GitHub username.' },
-        },
-        required: ['username'],
-      },
-    },
-  },
-  {
-    type: 'function',
-    function: {
-      name: 'list_actions',
-      description: 'Lists the GitHub actions running in a GitHub repository.',
-      parameters: {
-        type: 'object',
-        properties: {
-          username: { type: 'string', description: 'The GitHub username.' },
-          repoName: { type: 'string', description: 'The repository name.' },
-          status: { type: 'string', description: 'The status of the actions (optional). Defaults to in_progress if not provided' },
-        },
-        required: ['username', 'repoName'],
-      },
-    },
-  },
-  {
-    type: 'function',
-    function: {
-      name: 'list_branches',
-      description: 'Lists branches for a given GitHub repository.',
-      parameters: {
-        type: 'object',
-        properties: {
-          username: { type: 'string', description: 'The GitHub username.' },
-          repoName: { type: 'string', description: 'The repository name.' },
-        },
-        required: ['username', 'repoName'],
-      },
-    },
-  },
-  {
-    type: 'function',
-    function: {
-      name: 'list_commit_history',
-      description: 'Lists commit history for a file in a GitHub repository.',
-      parameters: {
-        type: 'object',
-        properties: {
-          username: { type: 'string', description: 'The GitHub username.' },
-          repoName: { type: 'string', description: 'The repository name.' },
-          filePath: { type: 'string', description: 'The file path.' },
-        },
-        required: ['username', 'repoName', 'filePath'],
-      },
-    },
-  },
-  {
-    type: 'function',
-    function: {
-      name: 'list_directory_contents',
-      description: 'Lists the contents of a directory in a GitHub repository.',
-      parameters: {
-        type: 'object',
-        properties: {
-          username: { type: 'string', description: 'The GitHub username.' },
-          repoName: { type: 'string', description: 'The repository name.' },
-          path: { type: 'string', description: 'The directory path (optional). Defaults to root if not provided' },
-        },
-        required: ['username', 'repoName'],
-      },
-    },
-  },
-];
-
-// Define the getFunctionDefinitionsForTool function
-function getFunctionDefinitionsForTool() {
-  return funcs;
-}
-
-function getAvailableFunctions() {
-  return availableFunctionsRegistry;
-}
-
 module.exports = {
   createGithubPullRequest,
-  getAvailableFunctions,
-  getFunctionDefinitionsForTool, // Renamed function
+  fetchRepoContentsRecursive,
   listBranches,
   listCommitHistory,
   listDirectoryContents,
